Lazily evaluate messages
This PR turns message generation in each method into a function, so that
it can be lazily evaluated only when needed. It's necessary to do it this
way because of the way 'not' works.

I'm not sure if there's a cleaner solution, but this yields a significant
performance increase as message generation on large objects with nested
properties is an expensive operation.